3.2.Database Migrations{Migration a Second Change}
Migration a Second Change
To see how nice this is, let's make one more change. Open the Article class. See the slug field?
This will eventually be used to identify the article in the URL. And so, this must be unique across every article in the table.
To guarantee that this is unique in the database, add unique=true
This option does only one thing: it tells Doctrine that it should create a unique index in the database for this column.