4.1.Saving Entities
Put on your publishing hat, because it's time to write some thoughtful space articles and insert some rows into our article table! And, good news! This is probably one of the easiest things to do in Doctrine.
Let's create a new controller called ArticleAdminController. We'll use this as a place to add new articles. Make it extend the normal AbstractController
And create a public function new()
Above, add the @Route() - make sure to auto-complete the one from Symfony Components so that PhpStorm adds the use statement. For the URL, how about /admin/article/new
We're not actually going to build a real page with a form here right now. Instead, I just want to write some code that saves a dummy article to the database.
But first, to make sure I haven't screwed anything up, return a new Response: the one from HttpFoundation with a message:

  space rocks... include comets, asteroids & meteoroids
Now, we should be able to find the browser and head to /admin/article/new. Great!