4.2.Saving Entities{Creating the Article Object}
Creating the Article Object
So, here's the big question: how do you save data to the database with Doctrine? The answer... is beautiful: just create an Article object with the data you need, then ask Doctrine to put it into the database.
Start with $article = new Article()
For this article's data, go back to the "Why Asteroids Taste like Bacon" article: we'll use this as our dummy news story. Copy the article's title, then call $article->setTitle() and paste
Oh, and the generator also made all the setter methods return $this, which means you can chain your calls, like: ->setSlug(), then copy the last part of the URL, and paste here. Oh, but we need to make sure this is unique... so just add a little random number at the end
Then, ->setContent(). And to get this, go back to ArticleController, copy all of that meaty markdown and paste here. Ah, make sure the content is completely not indented so the multi-line text works
Much better! The last field is publishedAt. To have more interesting data, let's only publish some articles. So, if a random number between 1 to 10 is greater than 2, publish the article: $article->setPublishedAt() with new \DateTime() and sprintf('-%d days') with a bit more randomness: 1 to 100 days old
Perfect! Now... stop. I want you to notice that all we've done is create an Article object and set data on it. This is normal, boring, PHP code: we're not using Doctrine at all yet. That's really cool.