5.1.Querying for Data!
Hey! There are rows in our article table! So let's update the news page to not show this hard-coded article, but instead to query the database and print real, dynamic data.
Open ArticleController and find the show() method. This renders that page. As I mentioned earlier, DoctrineBundle gives us one service - the EntityManager - that has the power to save and fetch data. Let's get it here: add another argument: EntityManagerInterface $em.
When you want to query for data, the first step is always the same: we need to get the repository for the entity: $repository = $em->getRepository() and then pass the entity class name: Article::class.
This repository object knows everything about how to query from the article table. We can use it to say $article = $repository->. Oh, nice! It has some built-in methods, like find() where you can pass the $id to fetch a single article. Or, findAll() to fetch all articles. With the findBy() method, you can fetch all articles where a field matches some value. And findOneBy() is the same, but only returns one Article. Let's use that: ->findOneBy() and pass it an array with 'slug' => $slug.
This will fetch one row where the slug field matches this value. These built-in find methods are nice... but they can't do much more than this. But, don't worry! We will of course learn how to write custom queries soon.
Above this line, just to help my editor, I'll tell it that this is an Article object.
And... hold on, that's important! When you query for something, Doctrine returns objects, not just an associative arrays with data. That's really the whole point of Doctrine! You need to stop thinking about inserting and selecting rows in a database. Instead, think about saving and fetching objects... almost as if you didn't know that a database was behind-the-scenes.