5.2.Querying for Data! {Handling 404's}
Handling 404's
At this point, it's possible that there is no article in the database with this slug. In that case, $article will be null. How should we handle that? Well, in the real world, this should trigger a 404 page. To do that, say if ! $article, then, throw $this->createNotFoundException(). Pass a descriptive message, like: No article for slug "%s" and pass $slug
I want to dig a little bit deeper to see how this work. Hold Command on a Mac - or Ctrl otherwise - and click this method. Ah, it comes from a trait that's used by the base AbstractController. Fascinating! It just throws an exception!
In Symfony, to trigger a 404, you just need to throw this very special exception class. That's why, in the controller, we throw $this->createNotFoundException(). The message can be as descriptive as possible because it will only be shown to you: the developer.
After all of this, let's dump() the $article to see what it looks like and die.
Head back to your browser and first, refresh. Ok! This is the 404 page: there's nothing in the database that matches this slug: all the real slugs have a random number at the end. We see the helpful error message because this is what the 404 page looks like for developers. But of course, when you switch into the prod environment, your users will see a different page that you can customize.
We're not going to talk about how to customize error pages... because it's super friendly and easy. Just Google for "Symfony customize error pages" and... have fun! You can create separate pages for 404 errors, 403 errors, 500 errors, or whatever your heart desires.
To find a real slug, go back to /admin/article/new. Copy that slug, go back, paste it and... it works! There is our full, beautiful, well-written, inspiring, Article object... with fake content about meat. Having an object is awesome! We are now... dangerous.