5.3.Querying for Data! {Rendering the Article Data: Twig Magic}
Rendering the Article Data: Twig Magic
Back in the controller, remove the dump(). Keep the hardcoded comments for now. But, remove the $articleContent. Let's also remove the markdown parsing code and the now-unused argument.
We'll process the markdown in the template in a minute: Back down at render(), instead of passing title, articleContent and slug, just pass article.
Now, open that template! With the Symfony plugin, you can cheat and hold Command or Ctrl and click to open it. Or, it's just in templates/article.
Updating the template is a dream. Instead of title, print article.title. Oh, and in many cases... but not always... you'll get auto-completion based on the methods on your entity class!
But look closely: it's auto-completing getTitle(). But when I hit tab, it just prints article.title. Behind the scenes, there is some serious Twig magic happening. When you say article.title, Twig first looks to see if the class has a title property. It does! But since that property is private, it can't use it. No worries! It then looks for a getTitle() method. And because that exists:
 93 lines  src/Entity/Article.php
class Article
{
    private $title;
    public function getTitle(): ?string
    {
        return $this->title;
    }
}
It calls it and prints that value.
This is really cool because our template code can be simple: Twig figures out what to do. If you were printing a boolean field, something like article.published, Twig would also look for isPublished() a hasPublished() methods. And, if article were an array, the dot syntax would just fetch the keys off of that array. Twig: you're the bomb.
Let's update a few more places: article.title, then, article.slug, and finally, for the content, article.content, but then |markdown.
The KnpMarkdownBundle gives us a markdown filter, so that we can just process it right here in the template.
Ready to try it? Move over, deep breath, refresh. Yes! It works! Hello dynamic title! Hello dynamic bacon content!

See your Queries in the Profiler

Oh, and I have a wonderful surprise! The web debug toolbar now has a database icon that tells us how many database queries this page executed and how long they took. But wait, there's more! Click the icon to go into the profiler. Yes! This actually lists every query. You can run "EXPLAIN" on each one or view a runnable query. I use this to help debug when a particularly complex query isn't returning the results I expect.
So, um, yea. This is awesome. Next, let's take a quick detour and have some fun by creating a custom Twig filter with a Twig extension. We need to do this, because our markdown processing is no longer being cached. Boo.