6.3.Fun with Twig Extensions! {Generating a Twig Extension}
We need a custom filter, so delete getFunctions() and then change the filter name to cached_markdown. Over on the right, this is the method that will be called when the user uses the filter. Let's call our method processMarkdown. Point to that from the filter.
To make sure things are working, for now, in processMarkdown(), just return strtoupper($value). Sweet! In the Twig template, use it: |cached_markdown.
Oh, and two important things. One, when you use a filter, the value to the left of the filter will become the first argument to your filter function. So, $value will be the article content in this case.
Second, check out this options array when we added the filter. This is optional. But when you say is_safe set to html
It tells Twig that the result of this filter should not be escaped through htmlentities(). And... that's perfect! Markdown gives HTML code, and so we definitely do not want that to be escaped. You won't need this option on most filters, but we do want it here.
And... yea. We're done! Thanks to Symfony's autoconfiguration system, our Twig extension should already be registered with the Twig. So, find your browser, high-five your dog or cat, and refresh!
It works! I mean, it's super ugly and angry-looking... but it works!