7.4.ago Filter with KnpTimeBundle{Using the ago Filter}
Using the ago Filter
Ok, let's use that filter! Back in the template, replace the date filter with |ago
That's it. Find the page, refresh and... perfect! 27 days ago. So much nicer!
Next, I want to talk a little bit more about the AppExtension Twig extension because, for a very subtle but important reason, it has a performance problem.