8.2.Service Subscriber: Lazy Performance{Creating a Service Subscriber}
Creating a Service Subscriber
We somehow want to tell Symfony to pass us the MarkdownHelper, but not actually instantiate it until, and unless, we need it. That's totally possible.
But, it's a little bit tricky until you see the whole thing put together. So, watch closely.
First, make your class implement a new interface: ServiceSubscriberInterface.
This will force us to have one new method. At the bottom of the class, I'll go to the "Code"->"Generate" menu - or Command+N on a Mac - and implement getSubscribedServices(). Return an array from this... but leave it empty for now.
Next, up on your constructor, remove the first argument and replace it with ContainerInterface - the one from Psr - $container. Also rename the property to $container.
At this point... if you're totally confused... no worries! Here's the deal: when you make a service implements ServiceSubscriberInterface, Symfony will suddenly try to pass a service container to your constructor. It does this by looking for an argument that's type-hinted with ContainerInterface. So, you can still have other arguments, as long as one has this type-hint.
But, one important thing: this  is not Symfony's big service container that holds hundreds of services. Nope, this is a mini-container, that holds a subset of those services. In fact, right now, it holds zero.