8.5.Service Subscriber: Lazy Performance{Understanding getSubscribedServices()}
If you just pass MarkdownHelper::class as the value, then that's also used as the key.
The end result is exactly the same as before, except MarkdownHelper is lazy! To prove it, put a die statement at the top of the MarkdownHelper constructor.
Now, go back to the article page and refresh. Not surprising: it hits the die statement when rendering the Twig template. But now, go back to the homepage. Yes! The whole page prints: MarkdownHelper is never instantiated.