9.1.All about Entity Repositories
With the article show page now dynamic, let's turn to the homepage... cause these news stories are totally still hardcoded. Open ArticleController and find the homepage() action
Perfect. Just like before, we need to query for the articles. This means that we need an EntityManagerInterface $em argument.
Next, we get the repository for the class: $repository = $em->getRepository(Article::class). And then we can say, $articles = $repository->findAll().
Nice! With this array of Article objects in hand, let's pass those into the template as a new articles variable
Now, to the template! Open homepage.html.twig and scroll down just a little bit. Yes: here is the article list.
Well, there's a "main" article on top, but I'm going to ignore that for now. Down below, add for article in articles with, at the bottom, endfor.
Then... just make things dynamic: article.slug, article.title, and for the three hours ago, if article.publishedAt is not null, print article.publishedAt|ago. If it's not published, do nothing. With this in place, delete the last two hardcoded articles.
Let's give it a try: find your browser and, refresh! Nice! You can see a mixture of published and unpublished articles. But... you can also see that the articles just printed out in whatever order they were created, regardless of the publish date. Space travellers demand fresh content! So let's print the newest articles first.