9.2.All about Entity Repositories{Controlling the ORDER BY}
Controlling the ORDER BY
Let's give it a try: find your browser and, refresh! Nice! You can see a mixture of published and unpublished articles. But... you can also see that the articles just printed out in whatever order they were created, regardless of the publish date. Space travellers demand fresh content! So let's print the newest articles first.
Head back to ArticleController. Hmm... the findAll() methods gives us everything... but it's pretty limited. In fact, it takes zero arguments: you can't control it at all.
But, some of the other methods are just a little bit more flexible. To control the order, use findBy() instead, pass this an empty array, and then another array with publishedAt set to DESC.
The first array is where you would normally pass some criteria for a WHERE clause. If we pass nothing, we get everything!

Try it - refresh! Much better!
Except... hmm... it probably does not make sense to show the unpublished articles on the homepage. And this is when things get a bit more interesting. Sure, you can pass simple criteria to findBy(), like slug equal to some value. But, in this case, we need a query that says WHERE publishedAt IS NOT NULL. That's just not possible with findBy()!