9.3.All about Entity Repositories{Hello ArticleRepository}
Hello ArticleRepository

And so... for the first time, we're going to write - drumroll - a custom query!
Let me show you something cool: when we originally generated our entity, the command created the Article class, but it also created an ArticleRepository class in the Repository directory. Try this: dump($repository) and, refresh. Guess what? This is an instance of that ArticleRepository!
Yes, there is a connection between the Article and ArticleRepository classes. In fact, that connection is explicitly configured right at the top of your Article class:

 93 lines  src/Entity/Article.php
/**
 * @ORM\Entity(repositoryClass="App\Repository\ArticleRepository")
 */
class Article
{
}
This says: when we ask for the Article class's repository, Doctrine should give us an instance of this ArticleRepository class:

 51 lines  src/Repository/ArticleRepository.php
use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
class ArticleRepository extends ServiceEntityRepository
{
}
Oh, and the built-in find*() methods actually come from one of the parent classes of ArticleRepository.

So... why the heck are we talking about this? Because, if you want to create a custom query, you can do that by creating a custom method inside of this class. And, hey! It even has a couple of examples