10.2.Custom Queries{Writing our Custom Query}
Writing our Custom Query
In our case, we won't need any arguments, and I'm going to simplify a bit. Let's say andWhere('a.publishedAt IS NOT NULL')
You can totally see how close this is to normal SQL. You can even put OR statements inside the string, like a.publishedAt IS NULL OR a.publishedAt > NOW().
Oh, and what the heck does the a mean? Think of this as the table alias for Article in the query - just like how you can say SELECT a.* FROM article AS a.
It could be anything: if you used article instead, you'd just need to change all the references from a. to article..
Let's also add our orderBy(), with a.publishedAt, DESC
Oh, and this is a good example of how we're referencing the property name on the entity. The column name in the database is actually published_at, but we don't use that here.
Finally, let's remove the max result.
Once you're done building your query, you always call getQuery() and then, to get the array of Article objects, getResult().
Below this method, there's an example of finding just one object.
It's almost the same: build the query, call getQuery(), but then finish with getOneOrNullResult().
So, in all normal situations, you always call getQuery(), then you'll either call getResult() to return many rows of articles, or getOneOrNullResult() to return a single Article object. Got it?
Now that our new findAllPublishedOrderedByNewest() method is done, let's go use it in the controller: $repository->, and there it is!
Let's give it a try! Move over and, refresh! Perfect! The order is correct and the unpublished articles are gone.