11.2.Query Logic Re-use & Shortcuts{Fancier Re-Use}
Reuse logic from previous commit is nice... but since I do this a lot, we can get a bit fancier. Create another private method called getOrCreateQueryBuilder() with a QueryBuilder argument like before, but make it optional
Here's the idea: when someone calls this method, if the query builder is passed, we'll just return it. Otherwise we will return a new one with $this->createQueryBuilder('a').
If you're not used to this syntax, it means that if a QueryBuilder object is passed, return that QueryBuilder object. If a QueryBuilder object is not passed, then create one.
This is cool, because now we can make the argument to addIsPublishedQueryBuilder() also optional.
Inside, use the new method: return $this->getOrCreateQueryBuilder() passing it $qb, and then our andWhere().
But the real beautiful thing is back up top. This whole method can now be one big chained call: return $this->addIsPublishedQueryBuilder() - and pass nothing.
It will create the QueryBuilder for us.
So not only do we have really nice public functions for fetching data, we also have some private functions to help us build our queries. Let's make sure it works. Find your browser and, refresh! It still looks good!